Changing from Int8 to UInt16
positionX and positionY and face_width are coded on 2 bytes, therefore they need to be coded on UInt16

+++ ros_opencv_converter/src/face_detection.cpp
@@ -10,7 +10,7 @@
 #include <opencv2/highgui/highgui.hpp>
 #include <opencv2/objdetect/objdetect.hpp>
 #include <opencv2/video/tracking.hpp>
-#include <std_msgs/Int8.h>
+#include <std_msgs/UInt16.h>
 
 #include <iostream>
 #include <fstream>
@@ -73,9 +73,9 @@ public:
 		// image_pub_ = it_.advertise("/face_detector/output_video", 1);
 
 		// Topic for Publishing the position of the face. Will buffer 2 messages before throwing away old ones
-		face_positionX_pub_ = nh_face.advertise<std_msgs::Int8>("/face_detector/face_position_X", 2);
-		face_positionY_pub_ = nh_face.advertise<std_msgs::Int8>("/face_detector/face_position_Y", 2);
-		face_width_pub_ = nh_face.advertise<std_msgs::Int8>("/face_detector/face_width", 2);
+		face_positionX_pub_ = nh_face.advertise<std_msgs::UInt16>("/face_detector/face_position_X", 2);
+		face_positionY_pub_ = nh_face.advertise<std_msgs::UInt16>("/face_detector/face_position_Y", 2);
+		face_width_pub_ = nh_face.advertise<std_msgs::UInt16>("/face_detector/face_width", 2);
 	}
 
 	~FaceDetector()
@@ -248,7 +248,7 @@ public:
 	{
 	    cv::rectangle(frame, face, Scalar(255,0,0));
 
-	    std_msgs::Int8 face_positionX, face_positionY, face_width;
+	    std_msgs::UInt16 face_positionX, face_positionY, face_width;
 	    face_positionX.data = face.x;
 	    face_positionY.data = face.y;
 	    face_width.data = face.width;
@@ -284,7 +284,7 @@ public:
 	    // Find initial face on screen...
 	    if (! found_face) 
 	    {
-	    	std_msgs::Int8 face_width;
+	    	std_msgs::UInt16 face_width;
 	    	face_width.data = 0;
 	    	face_width_pub_.publish(face_width);
 	        startMeasuringTime();
